+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+* **android:** place Yearly alarm dots on the jieqi/solstice ray and keep compact emojis visible
+
 ## [1.4.0](https://github.com/edwardlthompson/AstroAlarm/compare/v1.3.1...v1.4.0) (2026-09-02)